requirements.txt
blinker==1.7.0
click==8.1.7
dominate==2.9.1
Faker==25.2.0
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pygame==2.5.2
pysqlcipher3==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.10.0
visitor==0.1.3
Werkzeug==3.0.1
